refactor: change settingsnavv2 to v1
