+++ mobile-app/app/screens/AppNavigator/screens/Portfolio/screens/AddOrEditAddressBookScreenV2.tsx
@@ -15,7 +15,7 @@ import { openURL } from '@api/linking'
 import { translate } from '@translations'
 import { useCallback, useEffect, useLayoutEffect, useState } from 'react'
 import { useSelector } from 'react-redux'
-import { SettingsParamList } from '../../Settings/SettingsNavigatorV2'
+import { SettingsParamList } from '../../Settings/SettingsNavigator'
 import { WalletTextInputV2 } from '@components/WalletTextInputV2'
 import { Text, TouchableOpacity, View } from 'react-native'
 import { ButtonV2 } from '@components/ButtonV2'
