+++ mobile-app/app/screens/AppNavigator/screens/Settings/components/ResetButtonV2.tsx
@@ -4,7 +4,7 @@ import { translate } from '@translations'
 import { WalletAlert } from '@components/WalletAlert'
 import { View, Text } from 'react-native'
 import { NavigationProp, useNavigation } from '@react-navigation/native'
-import { SettingsParamList } from '../SettingsNavigatorV2'
+import { SettingsParamList } from '../SettingsNavigator'
 import React, { useCallback } from 'react'
 import { authentication, Authentication } from '@store/authentication'
 import { MnemonicStorage } from '@api/wallet/mnemonic_storage'
