+++ mobile-app/app/screens/AppNavigator/screens/Settings/screens/KnowledgeBaseScreenV2.tsx
@@ -9,7 +9,7 @@ import {
 import { StackScreenProps } from '@react-navigation/stack'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
-import { SettingsParamList } from '../SettingsNavigatorV2'
+import { SettingsParamList } from '../SettingsNavigator'
 
 type Props = StackScreenProps<SettingsParamList, 'KnowledgeBaseScreen'>
