+++ mobile-app/app/screens/AppNavigator/screens/Settings/screens/ServiceProviderScreenV2.tsx
@@ -10,7 +10,7 @@ import { MnemonicStorage } from '@api/wallet/mnemonic_storage'
 import { StackScreenProps } from '@react-navigation/stack'
 import { useAppDispatch } from '@hooks/useAppDispatch'
 import { useLogger } from '@shared-contexts/NativeLoggingProvider'
-import { SettingsParamList } from '../SettingsNavigatorV2'
+import { SettingsParamList } from '../SettingsNavigator'
 import { useServiceProviderContext } from '@contexts/StoreServiceProvider'
 import { useWalletNodeContext } from '@shared-contexts/WalletNodeProvider'
 import { WalletTextInputV2 } from '@components/WalletTextInputV2'
